# integration-hub/openapi.yaml
openapi: 3.0.3
info:
  title: Forge Integration Hub API
  version: 1.1.0-identity
  description: >
    Centralized Academy ↔ RMS synchronization and Forge Person Identity service.
    Academy and RMS must call POST /identity/resolve before creating people;
    personId is a Firestore auto-generated document id and is never client-editable.
servers:
  - url: /api/integrations/v1
security:
  - BearerAuth: []
  - IntegrationSecret: []
paths:
...
  /departments/{departmentId}:
    get:
      summary: Get department connection
      parameters:
        - $ref: "#/components/parameters/departmentId"
      responses:
        "200":
          description: Department
    patch:
      summary: Update department connection profile
      parameters:
        - $ref: "#/components/parameters/departmentId"
      responses:
        "200":
          description: Updated
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    IntegrationSecret:
      type: apiKey
      in: header
      name: X-Integration-Secret
  parameters:
    departmentId:
      name: departmentId
      in: path
      required: true
      schema:
        type: string
...